Move dynamic params to vm/evm and create contract_comm errors
