+++ params/protocol_params.go
@@ -116,7 +116,7 @@ var (
 	MinimumDifficulty      = big.NewInt(131072) // The minimum that the difficulty may ever be.
 	DurationLimit          = big.NewInt(13)     // The decision boundary on the blocktime duration used to determine whether difficulty should go up or not.
 
-	registrySmartContractAddress = common.HexToAddress("0x000000000000000000000000000000000000ce10")
+	RegistrySmartContractAddress = common.HexToAddress("0x000000000000000000000000000000000000ce10")
 )
 
 const (
